feat(shim-sev): Refactor exit() and exit_group()
Use normal sallyport syscall proxying for `exit()` and `exit_group()`
for all shims.
