+++ internal/syscall/src/lib.rs
@@ -83,12 +83,6 @@ pub trait SyscallHandler: AddressValidator + Sized {
     ) -> Result;
 
     /// syscall
-    fn exit(&mut self, status: libc::c_int) -> !;
-
-    /// syscall
-    fn exit_group(&mut self, status: libc::c_int) -> !;
-
-    /// syscall
     fn arch_prctl(&mut self, code: libc::c_int, addr: libc::c_ulong) -> Result;
 
     /// syscall
@@ -212,6 +206,31 @@ pub trait SyscallHandler: AddressValidator + Sized {
         }
 
         ret
+    }
+
+    /// Proxy an exit() syscall
+    fn exit(&mut self, status: libc::c_int) -> ! {
+        self.trace("exit", 1);
+
+        #[allow(unused_must_use)]
+        loop {
+            unsafe { self.proxy(request!(libc::SYS_exit => status)) };
+            self.attacked();
+        }
+    }
+
+    /// Proxy an exitgroup() syscall
+    ///
+    /// TODO: Currently we are only using one thread, so this will behave the
+    /// same way as exit(). In the future, this implementation will change.
+    fn exit_group(&mut self, status: libc::c_int) -> ! {
+        self.trace("exit_group", 1);
+
+        #[allow(unused_must_use)]
+        loop {
+            unsafe { self.proxy(request!(libc::SYS_exit_group => status)) };
+            self.attacked();
+        }
     }
 
     /// syscall
